Rebranded and various bug fixes
