+++ index.html
@@ -13,7 +13,7 @@
 
     <script src="/node_modules/@webcomponents/webcomponentsjs/webcomponents-loader.js"></script>
 
-    <script type="module" src="/src/showcase-application/showcase-application.js" crossorigin></script>
+    <script type="module" src="/src/mark-armendariz-application/mark-armendariz-application.js" crossorigin></script>
     <script src="./node_modules/web-animations-js/web-animations-next-lite.min.js"></script>
 
     <custom-style>
@@ -29,6 +29,6 @@
   
   </head>
   <body>
-    <showcase-application></showcase-application>
+    <mark-armendariz-application></mark-armendariz-application>
   </body>
 </html>
