--- src/component-library/page-component-library/sc-landing-page/sc-landing-page.js
@@ -1,120 +1,0 @@
-
-import {html, PolymerElement} from '@polymer/polymer/polymer-element.js';
-import {ScCard} from '../../sc-card/sc-card.js';
-import {} from '@polymer/polymer/lib/elements/dom-repeat';
-import {ScResumeEntry} from '../sc-resume-entry/sc-resume-entry.js';
-import '@polymer/polymer/lib/elements/dom-if.js';
-import {ScButton} from '../../sc-button/sc-button.js';
-/**
- * @customElement
- * @polymer
- */
-
-
-export class ScLandingPage extends PolymerElement {
-  
-  static get template() {
-    return html`
-      <style>
-        :host {
-          display: block;
-          background-image:url('src/images/landingPage.jpg');
-          background-attachment: fixed;
-          background-position: center;
-          background-repeat: no-repeat;
-          background-size: cover;
-          /*background-color: grey;*/
-
-        }
-        #landingFrame{
-          position: relative;
-          top: 50%;
-          left: 50%;
-          transform: translate(-50%, -50%);
-          font-size: 30pt;
-          text-align: center;
-          font-weight: 600;
-        }
-        #apptitle{
-          font-size: 40pt;
-          font-weight: 900;
-          color: var(--sc-main-theme-color);
-        }
-      </style>
-
-      <div id="landingFrame">
-        <div>Hi, I'm</div>
-        <div id='apptitle'>Mark Armendariz</div>
-        <sc-button text="Enter"></sc-button>
-      </div>
-
-    `;
-  }
-
-
-  static get properties() {
-    return {
-      /**
-       * Title of Resume Section
-       */
-      title:{
-        type: String
-      },
-
-      sectionId: {
-        type: String
-      },
-
-      hidden:{
-        type: Boolean,
-        reflectToAttribute: true,
-        value: false
-      },
-      route: {
-        type: String,
-        notify: true,
-        observer: "_checkRoute",
-        reflectToAttribute: true
-      },
-
-      selected:{
-        type: Boolean,
-        value: false,
-        notify: true
-      },
-
-      /**
-       * TODO: Replace with web service response
-       */
-      resumeSectionData:{
-        type: Object,
-        notify: true,
-        reflectToAttribute: true
-      }
-    };
-  }
-
-  ready(){
-    super.ready();
-    //this.$.frame.addEventListener('buttontap', e => this._handlePageButtonClick(e))
-  }
-
-  _handlePageButtonClick(e){
-    //this.selectedPage = e.detail;
-    console.log("home");
-    this.dispatchEvent(new CustomEvent('homeEnterTap', {bubbles: true, composed: true}));
-  }
-
-  _checkRoute(route){
-    if(route.path=="/"){
-      this.hidden = false;
-    }
-  }
-
-  hide(){
-    this.hidden=true;
-  }
-
-}
-
-window.customElements.define('sc-landing-page', ScLandingPage);
